backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1
Pillow==11.1.0
python-decouple==3.8
django-filter==23.5
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9; python_version < "3.13"
whitenoise==6.6.0
gunicorn==21.2.0












